+++ core/state_processor_xlayer.go
@@ -1,0 +1,32 @@
+// Copyright 2019 The go-ethereum Authors
+// This file is part of the go-ethereum library.
+//
+// The go-ethereum library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// The go-ethereum library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
+
+package core
+
+import (
+	"github.com/ledgerwatch/erigon/core/vm"
+	zktypes "github.com/ledgerwatch/erigon/zk/types"
+)
+
+func afterApplyTransaction(env vm.VMInterface, failed bool) []*zktypes.InnerTx {
+	innerTxs := env.GetInnerTxMeta().InnerTxs
+	if failed {
+		for _, innerTx := range innerTxs {
+			innerTx.IsError = true
+		}
+	}
+	return innerTxs
+}
