+++ core/state_transition.go
@@ -385,7 +385,7 @@ func (st *StateTransition) TransitionDb(refunds bool, gasBailout bool) (*Executi
 		ret   []byte
 		vmerr error // vm errors do not effect consensus and are therefore not assigned to err
 	)
-	//add InnerTx
+	//XLayer, add InnerTx
 	innerTx := &zktypes.InnerTx{
 		Dept:         *big.NewInt(0),
 		From:         sender.Address().String(),
@@ -403,11 +403,13 @@ func (st *StateTransition) TransitionDb(refunds bool, gasBailout bool) (*Executi
 		// It does get incremented inside the `Create` call, after the computation
 		// of the contract's address, but before the execution of the code.
 		ret, newAddr, st.gas, vmerr = st.evm.Deploy(sender, st.data, st.gas, st.value, intrinsicGas)
+		//XLayer, add InnerTx
 		innerTx.To = newAddr.String()
 	} else {
 		// Increment the nonce for the next transaction
 		st.state.SetNonce(msg.From(), st.state.GetNonce(sender.Address())+1)
 		ret, st.gas, vmerr = st.evm.Call(sender, st.to(), st.data, st.gas, st.value, bailout, intrinsicGas)
+		//XLayer, add InnerTx
 		innerTx.To = vm.AccountRef(*msg.To()).Address().String()
 	}
 	if refunds {
@@ -456,6 +458,7 @@ func (st *StateTransition) TransitionDb(refunds bool, gasBailout bool) (*Executi
 		)
 	}
 
+	// XLayer, add inner tx
 	innerTx.GasUsed = st.gasUsed()
 	return &ExecutionResult{
 		UsedGas:    st.gasUsed(),
