+++ eth/stagedsync/stage_mining_exec.go
@@ -92,7 +92,7 @@ func StageMiningExecCfg(
 // - resubmitAdjustCh - variable is not implemented
 func SpawnMiningExecStage(s *StageState, tx kv.RwTx, cfg MiningExecCfg, quit <-chan struct{}) error {
 	cfg.vmConfig.NoReceipts = false
-	cfg.vmConfig.NoInnerTxs = false
+	cfg.vmConfig.NoInnerTxs = false //XLayer operation
 	chainID, _ := uint256.FromBig(cfg.chainConfig.ChainID)
 	logPrefix := s.LogPrefix()
 	current := cfg.miningState.MiningBlock
