+++ ethdb/prune/storage_mode.go
@@ -157,6 +157,7 @@ func Get(db kv.Getter) (Mode, error) {
 		prune.CallTraces = blockAmount
 	}
 
+	// XLayer operations
 	blockAmount, err = get(db, kv.PruneReceipts)
 	if err != nil {
 		return prune, err
